# backend/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'e-Money API Documentation'
  description: 'Hệ thống REST API quản lý tài chính cá nhân e-Money'
  version: 1.0.0
servers:
  -
    url: /api
    description: 'Cổng API chính'
paths:
  /accounts:
    get:
      tags:
        - 'Ví & Tài khoản'
      summary: 'Lấy danh sách các loại ví'
      operationId: 96daed062b209799ebd8c7e84dc97aed
      responses:
        200:
          description: 'Dữ liệu danh sách ví'
        401:
          description: 'Chưa xác thực'
      security:
        -
          sanctum: []
    post:
      tags:
        - 'Ví & Tài khoản'
      summary: 'Tạo tài khoản/ví mới'
      operationId: b3234b57f04a5c23b5d70ac83678df8c
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  example: 'Ví Tiền mặt'
                icon:
                  type: string
                  example: Wallet
                balance:
                  type: number
                  format: float
                  example: 1000000
                is_default:
                  type: boolean
                  example: false
              type: object
      responses:
        201:
          description: 'Đã tạo ví thành công'
        422:
          description: 'Dữ liệu không hợp lệ'
      security:
        -
          sanctum: []
  /auth/register:
    post:
      tags:
        - 'Xác thực'
      summary: 'Đăng ký tài khoản mới'
      operationId: 08136088c4862c313b35d6518bab8d3a
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                full_name:
                  type: string
                  example: 'Nguyen Van A'
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: password123
              type: object
      responses:
        210:
          description: 'Đã đăng ký thành công'
        422:
          description: 'Dữ liệu không hợp lệ'
  /auth/login:
    post:
      tags:
        - 'Xác thực'
      summary: 'Đăng nhập'
      operationId: 99d66635c4992aeaa6aa44ff653d0563
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: password123
              type: object
      responses:
        200:
          description: 'Đăng nhập thành công và trả về Token'
        401:
          description: 'Thông tin đăng nhập không chính xác'
  /auth/me:
    get:
      tags:
        - 'Xác thực'
      summary: 'Lấy profile người dùng hiện tại'
      operationId: 7edae2cdbaafc10dff8c9c4198b320a2
      responses:
        200:
          description: 'Thông tin người dùng'
      security:
        -
          sanctum: []
  /auth/logout:
    post:
      tags:
        - 'Xác thực'
      summary: 'Đăng xuất'
      operationId: 6abf3b64a4bc7838d56346f05a5153af
      responses:
        200:
          description: 'Đã đăng xuất thành công'
      security:
        -
          sanctum: []
  /budgets:
    get:
      tags:
        - 'Ngân sách'
      summary: 'Lấy danh sách thiết lập ngân sách'
      operationId: 49a5c108bcc6c0f36d6721daf5ff8732
      responses:
        200:
          description: 'Danh sách ngân sách'
      security:
        -
          sanctum: []
    post:
      tags:
        - 'Ngân sách'
      summary: 'Thiết lập ngân sách chi tiêu'
      operationId: 99b08fee5d3d4d75f33713e9343f0c91
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                category_id:
                  type: integer
                  example: 1
                limit_amount:
                  type: number
                  format: float
                  example: 5000000
                month:
                  type: string
                  example: 2024-03
              type: object
      responses:
        200:
          description: 'Đã lưu thiết lập'
        422:
          description: 'Dữ liệu không hợp lệ'
      security:
        -
          sanctum: []
  '/budgets/{id}':
    delete:
      tags:
        - 'Ngân sách'
      summary: 'Xóa thiết lập ngân sách'
      operationId: 46c1727c8456d5ae253255b278b71ffd
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Đã xóa'
        404:
          description: 'Không tìm thấy'
      security:
        -
          sanctum: []
...
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Chỉ cần dán mã Token vào đây (không cần gõ chữ Bearer)'
...
security:
  -
    sanctum: []